Refactor selection calculations; draw visible AAs
This commit mainly updates `amino-acid-slider`, and tries to:

* Make it clear when we're working in percent and when in pixels
* Make it clearer how various calculations are derived, with fewer magic numbers
* Draw SVG only within the visible box, and calculate positions of AA images
* Always select the appropriate AA, given the selection box

This also removes the anti-pattern of using state to store derived properties,
and replaces it with getters that can be used by various methods. In theory
these could be memoized for efficiency, but it doesn't seem worth it.

With these changes, it's easier to share highlight location between the slider
and the protein, and easier to share selection index between the slider and
the info box.

The `Protein` view was changed to a stateless functional component, because there
is no need to save state after removing the derived-props anti-pattern.