Allow user to select AAs outside selection box
For this, we also have to make the x location of the selected AA explicit,
which adds some bookkeeping, but, as a plus, removes some magic code from
the infobox.

Update the click handler to allow clicking on different AAs, and to make
clicking on the same one the toggle.

[#158988016]