Replace allele string with DNA sequence
This allows (requires) proteins to be defined with strings of DNA
instead of directly as amino acid chains.

This also replaces codon-utils and amino-acid-types with a more generic
and broader amino-acid-utils, which can do the codon lookups and contains
the database of properties.

The amino acid also contains the colors, which is irrelevant for this
commit but will be used later.

[#158988040]